fastapi1/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
docopt==0.6.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pipreqs==0.4.13
psycopg2==2.9.6
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
yarg==0.1.9
